refactor: Move RestaurantSearchResponse to dto.response package
